added spotltl header and libs to ltl-compare

+++ src/ltl-compare/include/ltl-compare.hh
@@ -1,4 +1,4 @@
 
 #include <string>
 //spot includes
-int semanticComparison(std::string a1, std::string a2);
+double semanticComparison(std::string a1, std::string a2);
